Moved map into subcomponent
